+++ sdk/storage/azure_storage_blob/src/clients/blob_client.rs
@@ -10,7 +10,7 @@ use crate::{
         http_ranges::IntoRangeHeader, BlobClientDownloadOptions, BlobClientDownloadResult,
         BlobClientUploadOptions, BlobClientUploadResult, StorageErrorCode,
     },
-    partitioned_transfer::{self, PartitionedDownloadBehavior},
+    partitioned_transfer::{self, defaults, PartitionedDownloadBehavior},
     AppendBlobClient, BlockBlobClient, PageBlobClient,
 };
 use async_trait::async_trait;
@@ -24,7 +24,7 @@ use azure_core::{
     },
     tracing, Bytes, Result,
 };
-use std::{num::NonZero, ops::Range, sync::Arc};
+use std::{ops::Range, sync::Arc};
 
 impl BlobClient {
     /// Creates a new BlobClient, using Entra ID authentication.
@@ -205,10 +205,13 @@ impl BlobClient {
         options: Option<BlobClientDownloadOptions<'_>>,
     ) -> Result<BlobClientDownloadResult> {
         let options = options.unwrap_or_default();
-        let parallel = options.parallel.unwrap_or(DEFAULT_DOWNLOAD_PARALLEL);
+        let parallel = options
+            .parallel
+            .unwrap_or_else(defaults::default_concurrency);
         let partition_size = options
             .partition_size
-            .unwrap_or(DEFAULT_DOWNLOAD_PARTITION_SIZE);
+            .map(defaults::clamp_download_partition_size)
+            .unwrap_or_else(defaults::default_download_partition_size);
         // Construct exhaustively to catch new options.
         let get_range_options = BlobClientDownloadInternalOptions {
             encryption_algorithm: options.encryption_algorithm,
@@ -288,10 +291,6 @@ impl BlobClient {
         }
     }
 }
-
-// unwrap evaluated at compile time
-const DEFAULT_DOWNLOAD_PARALLEL: NonZero<usize> = NonZero::new(4).unwrap();
-const DEFAULT_DOWNLOAD_PARTITION_SIZE: NonZero<usize> = NonZero::new(4 * 1024 * 1024).unwrap();
 
 struct BlobClientDownloadBehavior<'a> {
     client: GeneratedBlobClient,