+++ sdk/storage/azure_storage_blob/src/partitioned_transfer/mod.rs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+pub(crate) mod defaults;
 mod download;
 mod upload;